requirements.txt
Flask==1.1.2
Requests==2.31.0
SQLAlchemy==2.0.15
mysqlclient~=2.1.1
python-dotenv~=1.0.0
gunicorn==20.1.0
Jinja2==2.11.2
PyMySQL==1.0.3
MarkupSafe==1.1.1
mysql==0.0.3
mysql-connector==2.2.9
itsdangerous==1.1.0
Werkzeug==1.0.1
